change ReportManager Aging Report which send the list of InvioceEntity

+++ Core/Data/Entities/CustomerEntity.cs
@@ -23,11 +23,11 @@ namespace Core.Data.Entities {
 
         [DataType(DataType.Currency)]
         [Column(TypeName = "decimal(18, 2)")]
-        public double? CreditLimit { get; set; }
+        public decimal? CreditLimit { get; set; }
 
         [DataType(DataType.Currency)]
         [Column(TypeName = "decimal(18, 2)")]
-        public double? CreditUtilized { get; set; }
+        public decimal? CreditUtilized { get; set; }
 
         [ForeignKey("Address")]
         [Column("CustomerAddress_Id")]
